backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.2
djangorestframework==3.15.1
imbalanced-learn==0.12.2
joblib==1.4.2
numpy==1.26.4
python-dateutil==2.9.0.post0
scikit-learn==1.4.2
scipy==1.14.0
sqlparse==0.5.1
boto3==1.35.24
psycopg2-binary==2.9.9
dj-database-url==2.2.0
gunicorn==23.0.0
threadpoolctl==3.5.0
tzdata==2025.1
xgboost==2.0.3
whitenoise==6.9.0
python-dotenv==1.1.0